Factored out components that support Help.js into their own modules, replaced paragraph with 'Copy URL' text.
