+++ website/js/Help.js
@@ -5,8 +5,8 @@ import classNames from "classnames";
 import update from 'immutability-helper';
 import BetaTag from "./components/BetaTag";
 import {debounce} from "lodash";
-import Mark from 'mark.js';
-import {idHelpClicked} from "./util";
+import SearchController, {EXTRA_SEARCH_PADDING} from "./components/help/SearchController";
+import HardlinkHelper, {idHelpClicked} from "./components/help/HardlinkHelper";
 // import debounce from 'lodash/debounce';
 const React = require('react');
 const RawHTML = require('./RawHTML');
@@ -17,185 +17,7 @@ var $ = require('jquery');
 const slugify = require('./slugify');
 const content = require('./content');
 
-const navbarHeight = 70;
-// extra padding when scrolling to a search result
-const EXTRA_SEARCH_PADDING = 8;
-
-/**
- * The SearchController renders as a sticky search box with extra controls for navigating the search results.
- *
- * When beginning a search, all collapsible elements in the targeted element's tree will be collapsed. Strings that
- * match the query will be wrapped in <span class="highlighted"> tags, and all ancestors of the string that are
- * collapsible will be expanded.
- *
- * When the site's mode is toggled, search results will be cleared since they won't be valid for the new text anyway.
- */
-const SearchController = React.createClass({
-    getInitialState() {
-        return {
-            searchTerm: '',
-            searching: false,
-            matched: 0,
-            currentMark: null
-        };
-    },
-
-    searchChanged(e) {
-        this.setState({
-            searchTerm: e.target.value,
-            searching: true
-        }, () => {
-            this.debouncedSearchResponse();
-        });
-    },
-
-    searchResponse() {
-        if (!this.state.searchTerm || this.state.searchTerm === '') {
-            // remove any marks if they cleared the search
-            this.setState({
-                matched: 0,
-                currentMark: null,
-                searching: false
-            }, () => {
-                this.searcher.unmark();
-
-                $(this.props.target).find('*[data-expander-id]').each((idx, elem) => {
-                    this.props.setExpansion($(elem).data('expander-id'), false);
-                });
-            });
-            return;
-        }
-
-        // perform full matching against 'target'
-        // (first we unmark, then mark, then deal with the match results)
-        this.searcher.unmark({
-            done: () => {
-                this.pendingUpdate = new Set();
-
-                // then iteratively expand while searching for marks
-                this.searcher.mark(this.state.searchTerm, {
-                    element: 'span',
-                    className: 'highlighted',
-                    done: (totalMarks) => {
-                        this.setState({
-                            searching: false,
-                            currentMark: null,
-                            matched: totalMarks
-                        });
-
-                        // set all the elements that can be toggled to their match status
-                        $('*[data-expander-id]').each((idx, elem) => {
-                            const targetID = $(elem).data('expander-id');
-                            this.props.setExpansion(targetID, this.pendingUpdate.has(targetID));
-                        });
-                    },
-                    each: (elem) => {
-                        const me = this;
-
-                        // check if it has ancestors that need to be expanded and add them to the expanded list
-                        $(elem).click(function() {
-                            const $highlightSet = $('.highlighted').removeClass("focused");
-
-                            // set this element to the currently-selected index
-                            $(this).addClass("focused");
-                            me.setState({ currentMark: $highlightSet.index(this) });
-                        })
-                        .parents('*[data-expander-id]').each((idx, elem) => {
-                            this.pendingUpdate.add($(elem).data('expander-id'));
-                        });
-                    }
-                });
-            }
-        });
-    },
-
-    componentWillReceiveProps(nextProps) {
-        if (nextProps.researchMode !== this.props.researchMode) {
-            // reinitialize if they've switched help page modes
-            this.setState({
-                searchTerm: '',
-                searching: false,
-                matched: 0,
-                currentMark: null
-            });
-
-            // clear any marks and recreate the marker
-            this.searcher.unmark({
-                done: () => {
-                    this.searcher = new Mark(this.props.target);
-                }
-            });
-        }
-    },
-
-    componentWillMount() {
-        this.searcher = new Mark(this.props.target);
-        this.debouncedSearchResponse = debounce(this.searchResponse, 300);
-
-        this.navForward = this.navMarks.bind(this, true);
-        this.navBackward = this.navMarks.bind(this, false);
-    },
-
-    navMarks(forward) {
-        const $highlightSet = $('.highlighted').removeClass("focused");
-        let nextMark = this.state.currentMark;
-
-        if (nextMark === null) {
-            // initialize currentMark if we haven't navigated anything previously
-            nextMark = forward ? 0 : $highlightSet.length - 1;
-        } else {
-            // apply navigation
-            nextMark = (nextMark + (forward ? 1 : -1)) % $highlightSet.length;
-            if (nextMark < 0) {
-                nextMark = $highlightSet.length + nextMark;
-            }
-        }
-
-        this.setState({
-            currentMark: nextMark
-        }, () => {
-            const $targetElem = $($highlightSet.get(this.state.currentMark)).addClass("focused");
-            // move to whatever we navigated to
-            window.scrollTo({
-                // we want the element to not be covered by the navbar or the sticky search header, with some
-                // extra cosmetic padding, EXTRA_PADDING, past the header as well
-                top: $targetElem.offset().top - navbarHeight - (this.props.headerElem.outerHeight() + EXTRA_SEARCH_PADDING),
-                behavior: 'smooth'
-            });
-        });
-    },
-
-    render() {
-        return (
-            <div className="input-group has-feedback has-search">
-                <div className="input-group-addon">
-                    <span className={`glyphicon ${this.state.searching ? "glyphicon-refresh glyphicon-spin" : "glyphicon-search"}`} />
-                </div>
-                <input type="text" className="form-control" placeholder="Search" value={this.state.searchTerm} onChange={this.searchChanged} />
-                {
-                    (this.state.matched > 0) && (
-                        <span className="input-group-addon">
-                         { this.state.currentMark !== null && `${this.state.currentMark + 1} / ` }
-                            { this.state.matched}
-                        </span>
-                    )
-                }
-                <div className="input-group-btn">
-                    <button type="button" disabled={this.state.matched <= 0} onClick={this.navForward} className="btn btn-default">
-                        <span className="glyphicon glyphicon-triangle-bottom" />
-                    </button>
-                    <button type="button" disabled={this.state.matched <= 0} onClick={this.navBackward} className="btn btn-default">
-                        <span className="glyphicon glyphicon-triangle-top" />
-                    </button>
-                </div>
-            </div>
-        );
-    }
-});
-SearchController.propTypes = {
-    target: React.PropTypes.string.isRequired,
-    setExpansion: React.PropTypes.func.isRequired
-};
+export const navbarHeight = 70;
 
 const CollapsableListItem = React.createClass({
     mixins: [State, BootstrapMixin, CollapsableMixin],
@@ -232,7 +54,7 @@ const CollapsableListItem = React.createClass({
                     <small><Glyphicon  glyph={this.props.expanded ? "chevron-down" : "chevron-right"} /> </small>
                     <span style={{verticalAlign: "text-bottom"}}>{header}</span>
                 </a>
-                <a className="id-helper-tip" onClick={(e) => { idHelpClicked(e); }} href={`#${id}`}>&para;</a>
+                <HardlinkHelper id={id} />
             </h4>
         );
 
@@ -298,7 +120,13 @@ const Help = React.createClass({
             setTimeout(function () {
                 var el = document.getElementById(fragment);
                 if (el) {
-                    window.scrollTo(0, el.getBoundingClientRect().top - navbarHeight);
+                    const elemTop = $(el).offset().top;
+                    const headerOffset = navbarHeight + $('.header-sticky').outerHeight() + EXTRA_SEARCH_PADDING;
+                    console.log("Element top: ", elemTop, "; header offset: ", headerOffset, "; total: ", elemTop - headerOffset);
+                    window.scrollTo({
+                        top: elemTop - headerOffset,
+                        behavior: 'smooth'
+                    });
                 }
             }, 0);
         }
@@ -368,7 +196,7 @@ const Help = React.createClass({
                 const actualId = id ? id : slugify(name);
 
                 let header = [<span key="header_name" className="identifier" id={actualId}>{name}</span>];
-                header.push(<a  key="header_id_helper" className="id-helper-tip" href={`#${actualId}`}>&para;</a>);
+                header.push(<HardlinkHelper id={actualId} />);
 
                 let body = [];
                 if (contents) {
@@ -431,8 +259,11 @@ const Help = React.createClass({
 
                 <Row ref={(me) => { if (me) { this.headerElem = $(me.getDOMNode()); } }} className="header-sticky">
                     <Col smOffset={1} sm={10} className="help-search-header">
-                        {/*<h1>BRCA Exchange: Help</h1>*/}
-                        <SearchController researchMode={localStorage.getItem('research-mode')} setExpansion={this.setExpansion} headerElem={this.headerElem} target="#help-body" />
+                        <SearchController
+                            researchMode={localStorage.getItem('research-mode')}
+                            setExpansion={this.setExpansion}
+                            headerElem={this.headerElem} target="#help-body"
+                        />
                     </Col>
                 </Row>
