+++ website/js/RawHTML.js
@@ -1,9 +1,10 @@
 /*global module: false, require: false */
 'use strict';
 
+import {HARDLINK_GLYPH, idHelpClicked} from "./components/help/HardlinkHelper";
+
 const React = require('react');
 const $ = require('jquery');
-import {idHelpClicked} from "./util";
 
 const RawHTML = React.createClass({
     componentDidMount() {
@@ -22,10 +23,13 @@ const RawHTML = React.createClass({
             // add tooltips to anything that has an id
             const $idElems = $("*[id]", this.refs.me.getDOMNode());
             $idElems.each((idx, elem) => {
+                // TODO: ideally we'd replace this ad-hoc html element with the actual HardlinkHelper, although that
+                //  would be tricky/hacky since the contents of RawHTML aren't part of the react component tree...
+                //  (for now, it resembles the actual version enough that it shouldn't be a problem.)
                 $(elem).addClass("id-associated");
                 $("<a>")
                     .addClass("id-helper-tip")
-                    .html("&para;")
+                    .html(HARDLINK_GLYPH)
                     .click((e) => { idHelpClicked(e); })
                     .attr("href", `#${$(elem).attr("id")}`)
                     .appendTo(elem);
